Move branch and workspace related types to `ui` modules.
This is to indicate that they truly are only relevant for the user-interface,
with plenty of redundant information made to make it easier to consume in the frontend.
