+++ crates/but-cli/Cargo.toml
@@ -37,3 +37,4 @@ tracing-subscriber.workspace = true
 tracing.workspace = true
 dirs-next = "2.0.0"
 serde_json = "1.0.140"
+serde.workspace = true
