+++ crates/but-hunk-dependency/tests/hunk_dependency/ui.rs
@@ -296,7 +296,7 @@ mod util {
         pub repo: gix::Repository,
         /// All the stacks in the workspace
         #[allow(unused)]
-        pub stacks_entries: Vec<but_workspace::StackEntry>,
+        pub stacks_entries: Vec<but_workspace::ui::StackEntry>,
         /// The storage directory for GitButler.
         pub gitbutler_dir: PathBuf,
     }
@@ -320,7 +320,7 @@ mod util {
     impl TestContext {
         /// Find a stack which contains a branch with the given `short_name`.
         #[allow(unused)]
-        pub fn stack_with_branch(&self, short_name: &str) -> &but_workspace::StackEntry {
+        pub fn stack_with_branch(&self, short_name: &str) -> &but_workspace::ui::StackEntry {
             self.stacks_entries
                 .iter()
                 .find(|s| s.name() == Some(short_name.into()))
